+++ prisma/migrations/20251117031833_remove_delivery_schedule_from_purchase_option/migration.sql
@@ -1,0 +1,8 @@
+/*
+  Warnings:
+
+  - You are about to drop the column `deliverySchedule` on the `PurchaseOption` table. All the data in the column will be lost.
+
+*/
+-- AlterTable
+ALTER TABLE "PurchaseOption" DROP COLUMN "deliverySchedule";